Test cases written sorted among, and uncommented tests which are not working or missing
