+++ tests/test_manju.py
@@ -1,4 +1,3 @@
-from framework.selenium_plus import *
 from pages.flows.common import *
 
 
@@ -13,7 +12,7 @@ def my_fixture():
     # setup_stuff
     yield
     try:
-        # quit_driver()
+        quit_driver()
         pass
     finally:
         pass
@@ -40,121 +39,131 @@ def test_logout(login_citizen):
     logout()
 
 
-def test_my_complaints():
+def test_rate_closed_compalint():
+    # Create complaint from citizen to last mile employee verification
+    pgr_details = complaint_workflow_from_citizen_to_employee(complaint_action="Mark as Resolve")
+
+    # Complaint Details (Test Data)
+    complaint_number = pgr_details["complaint_number"]
+    status = pgr_details["status"]
+    print("Complaint Number:{0}, Complaint Status:{1}".format(complaint_number, status))
+    current_status = pgr_details["current_status"]
+
+    # Verification
+    print("Complaint Number:{0}, Complaint Status:{1}".format(complaint_number, current_status))
+
+    # rate the closed complaint
     citizen_login()
-    complaints = MyComplaintsPage()
-    # complaints.navigate()
-
-    cards = complaints.get_all_complaints()
-    card = cards[2]
-    card.track_complaint()
-
-
-def test_language_selection():
-    language_selection = LanguageSelectionPage()
-    language_selection.navigate()
-    language_selection.language("punjabi").language("hindi").language("english").submit()
+    rate_complaint = rate_closed_complaint(complaint_number)
+    status_before = rate_complaint["status_before_rate"]
+    status_after = rate_complaint["status_after_rate"]
+    print("Complaint Number:{0}, Complaint Status:{1}".format(complaint_number, status_before))
+    print("Complaint Number:{0}, Complaint Status:{1}".format(complaint_number, status_after))
 
 
-def test_user_registration():
-    user_reg = RegistrationPage()
-    user_reg.navigate()
-    user_reg.set("9988776655", "FirstName", "Bathinda").submit()
+def test_reopen_closed_complaint():
+    # Create complaint from citizen to last mile employee verification
+    pgr_details = complaint_workflow_from_citizen_to_employee(complaint_action="Mark as Resolve")
 
+    # Complaint Details (Test Data)
+    complaint_number = pgr_details["complaint_number"]
+    status = pgr_details["status"]
+    print("Complaint Number:{0}, Complaint Status:{1}".format(complaint_number, status))
+    current_status = pgr_details["current_status"]
 
-def test_profile_update(upload_photo=PROFILE_IMAGELIST):
+    # Verification
+    print("Complaint Number:{0}, Complaint Status:{1}".format(complaint_number, status))
+
+    # Reopen the closed complaint
     citizen_login()
-    TopMenuNavigationComponent().ham()
-    LoginPage().profile()
-    ProfilePage().update("Manjunatha S", "manju@ulb.in")
-    ProfilePage().change_profile_picture(upload_photo)
-    ProfilePage().save()
-    assert ProfilePage().toaster_message() == "Profile is Successfully Updated"
-    navigation = TopMenuNavigationComponent()
-    navigation.back()
-    logout()
+    reopen_complaint = reopen_closed_complaint(complaint_number)
+    status_before = reopen_complaint["status_before_reopen"]
+    status_after = reopen_complaint["status_after_reopen"]
+    print("Complaint Number:{0}, Complaint Status:{1}".format(complaint_number, status_before))
+    print("Complaint Number:{0}, Complaint Status:{1}".format(complaint_number, status_after))
 
 
-def test_discard_changes_in_profile():
+def test_reassign_complaint():
+    # Create complaint from citizen to last mile employee verification
+    pgr_details = complaint_workflow_from_citizen_to_employee(complaint_action="Request for Re-Assign")
+
+    # Complaint Details (Test Data)
+    complaint_number = pgr_details["complaint_number"]
+    status = pgr_details["status"]
+    print("Complaint Number:{0}, Complaint Status:{1}".format(complaint_number, status))
+    current_status = pgr_details["current_status"]
+
+    # Verification
+    print("Complaint Number:{0}, Complaint Status:{1}".format(complaint_number, current_status))
+
+    # Re-Assign complaint
+    gro_employee_login("Amardeep", "12345678")
+    reassign = complaint_reassign(complaint_number)
+    current_status = reassign["status"]
+    print("Complaint Number:{0}, Complaint Status:{1}".format(complaint_number, current_status))
+    # last_mile_employee_login("MamataDevi", "12345678")
+    # last_mile_employee_verification(complaint_number, "Mark as Resolve")
+
+
+def test_reject_complaint():
+    # Create complaint from citizen
     citizen_login()
-    TopMenuNavigationComponent().ham()
-    LoginPage().profile()
-    ProfilePage().update("Sathish", "sathish@ulb.in")
-
-
-def test_homepage():
-    citizen_login()
-    home_page = HomePage()
-    home_page.new_complaint()
-    home_page.navigate().click_my_complaint()
-
-
-def test_add_complaint(citizen_login, upload_photo=DEFAULT_IMAGELIST_THREE):
-    # Create a new complaint
-    create_new_complaint("Garbage", "Amritsar, Punjab, India ", "Street end", "Leakage of water", upload_photo)
-    # Acknowledgement on successful complaint submission
-    complaint_no = complaint_registration_number_recevied()
-    print(complaint_no)
-    # Search and view complaint created on My Complaints
-    open_complaint(complaint_no)
-    comment_on_complaint("Add comments")
-    # Navigate to the home page and logout
+    upload_photo = DEFAULT_IMAGELIST_TWO
+    complaint_info = create_new_complaint("No water or electricity in Public Toilet", "Amritsar, Punjab, India ",
+                                          "Street end",
+                                          "Leakage of water", upload_photo)
+    complaint_number = complaint_info["complaint_number"]
+    status = complaint_info["status"]
+    print("Complaint Number:{0}, Complaint Status:{1}".format(complaint_number, status))
     navigation = TopMenuNavigationComponent()
     navigation.back().back()
     logout()
+    quit_driver()
+
+    # GRO verification (Reject Complaint)
+    gro_employee_login("Amardeep", "12345678")
+    complaint_reject(complaint_number)
+    logout()
+    quit_driver()
+
+    # Citizen Login
+    citizen_login()
+    open_complaint(complaint_number)
+    status = get_current_status()
+    print("Complaint Number:{0}, Complaint Status:{1}".format(complaint_number, status))
+    navigation.back().back()
+    logout()
+    quit_driver()
 
 
-def test_pgr_workflow(citizen_login, upload_photo=DEFAULT_IMAGELIST_THREE):
-    # Create a new complaint
-    create_new_complaint("Garbage", "Amritsar, Punjab, India ", "Street end", "Leakage of water", upload_photo)
-    # Acknowledgement on successful complaint submission
-    complaint_no = complaint_registration_number_recevied()
-    print(complaint_no)
-    # Search and view complaint created on My Complaints
-    view_my_complaints(complaint_no)
-    comment_on_complaint("Add comments")
-    # Navigate to the home page and logout
+def test_reopen_rejected_complaint():
+    # Create complaint from citizen
+    citizen_login()
+    upload_photo = DEFAULT_IMAGELIST_TWO
+    complaint_info = create_new_complaint("No water or electricity in Public Toilet", "Amritsar, Punjab, India ",
+                                          "Street end",
+                                          "Leakage of water", upload_photo)
+    complaint_number = complaint_info["complaint_number"]
+    status = complaint_info["status"]
+    print("Complaint Number:{0}, Complaint Status:{1}".format(complaint_number, status))
     navigation = TopMenuNavigationComponent()
     navigation.back().back()
     logout()
     quit_driver()
-    # Login as GRO
+
+    # GRO verification (Reject Complaint)
     gro_employee_login("Amardeep", "12345678")
-    view_my_complaints(complaint_no)
-    assign_open_complaint(complaint_no, "AntrikshKumar")
-    logout()
-    quit_driver()
-    # Login as Last Mile Employee
-    last_mile_employee_login("Antarikshkumar", "12345678")
-    view_my_complaints(complaint_no)
-    resolve_assigned_complaint(complaint_no)
+    complaint_reject(complaint_number)
     logout()
     quit_driver()
 
-
-def test_view_my_complaint():
-    gro_employee_login("Amardeep", "12345678")
-    view_my_complaints("18/05/2018/000800")
-    assign_open_complaint("18/05/2018/000800", "V Sudheer")
-
-
-def test_citizen_should_file_complaint_with_one_image(citizen_login, upload_photo=DEFAULT_IMAGELIST_ONE):
-    # Create a new complaint
-    complaint_type = "Garbage"
-    location = "Amritsar, Punjab, India "
-    landmark = "Street end"
-    additional_details = "Uneven"
-
-    create_new_complaint(complaint_type, location, landmark, additional_details, upload_photo)
-
-
-def test_citizen_should_file_complaint_with_two_image(citizen_login):
-    test_citizen_should_file_complaint_with_one_image(citizen_login, upload_photo=DEFAULT_IMAGELIST_TWO)
-
-
-def test_citizen_should_file_complaint_with_three_image(citizen_login):
-    test_citizen_should_file_complaint_with_one_image(citizen_login, upload_photo=DEFAULT_IMAGELIST_THREE)
-
-
-def test_citizen_should_file_complaint_without_image(citizen_login):
-    test_citizen_should_file_complaint_with_one_image(citizen_login, [])
+    # Citizen Login: Reopen Rejected Complaint
+    citizen_login()
+    reopen_complaint = reopen_closed_complaint(complaint_number)
+    status_before = reopen_complaint["status_before_reopen"]
+    status_after = reopen_complaint["status_after_reopen"]
+    print("Complaint Number:{0}, Complaint Status:{1}".format(complaint_number, status_before))
+    print("Complaint Number:{0}, Complaint Status:{1}".format(complaint_number, status_after))
+    navigation.back().back()
+    logout()
+    quit_driver()
